[MOVE] 랭킹도메인 관련 모듈 랭킹 서버 하위로 이동
