+++ app/ranking/src/main/java/com/chinda/ranking/RankingApplication.java
@@ -1,0 +1,13 @@
+package com.chinda.ranking;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class RankingApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(RankingApplication.class, args);
+	}
+
+}
