+++ app/ranking/src/main/java/com/chinda/ranking/domain/record/CertificateRepository.java
@@ -1,6 +1,5 @@
-package com.core.core.domain.record.repository;
+package com.chinda.ranking.domain.record;
 
-import com.core.core.domain.record.Certificate;
 import org.springframework.data.repository.Repository;
 
 public interface CertificateRepository extends Repository<Certificate, Long> {
