+++ app/ranking/src/main/java/com/chinda/ranking/domain/record/RecordRepository.java
@@ -1,0 +1,6 @@
+package com.chinda.ranking.domain.record;
+
+import org.springframework.data.repository.Repository;
+
+public interface RecordRepository extends Repository<java.lang.Record, Long> {
+}
