+++ app/ranking/src/test/java/com/chinda/ranking/RankingApplicationTests.java
@@ -1,0 +1,13 @@
+package com.chinda.ranking;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class RankingApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
