--- core/src/main/java/com/core/core/domain/record/repository/RecordRepository.java
@@ -1,7 +1,0 @@
-package com.core.core.domain.record.repository;
-
-import com.core.core.domain.record.Record;
-import org.springframework.data.repository.Repository;
-
-public interface RecordRepository extends Repository<Record, Long> {
-}
